chore: unify migration config
